--- loadercomplex-fabric-12/src/main/java/com/enderzombi102/loaderComplex/fabric12/imixin/IItemMixin.java
@@ -1,10 +1,0 @@
-package com.enderzombi102.loaderComplex.fabric12.imixin;
-
-import net.minecraft.block.Block;
-import net.minecraft.item.Item;
-
-import java.util.Map;
-
-public interface IItemMixin {
-	Map<Block, Item> lc$getBlockItemRegistry();
-}
