--- loadercomplex-fabric-12/src/main/java/com/enderzombi102/loaderComplex/fabric12/impl/FabricServer.java
@@ -1,17 +1,0 @@
-package com.enderzombi102.loaderComplex.fabric12.impl;
-
-import com.enderzombi102.loadercomplex.api.utils.Server;
-import net.minecraft.server.MinecraftServer;
-
-public class FabricServer implements Server {
-	private final MinecraftServer backingServer;
-
-	public FabricServer(MinecraftServer server) {
-		this.backingServer = server;
-	}
-
-	@Override
-	public Object getObject() {
-		return this.backingServer;
-	}
-}
