+++ loadercomplex-fabric-12/src/main/java/com/enderzombi102/loadercomplex/fabric12/LoaderComplexFabric.java
@@ -1,0 +1,30 @@
+package com.enderzombi102.loadercomplex.fabric12;
+
+import com.enderzombi102.loadercomplex.fabric12.impl.FabricLoader;
+import com.enderzombi102.loadercomplex.LoaderComplex;
+import com.enderzombi102.loadercomplex.api.Loader;
+import net.fabricmc.api.ModInitializer;
+
+import java.util.ArrayList;
+
+public class LoaderComplexFabric extends LoaderComplex implements ModInitializer {
+	public static final ArrayList<FabricResourcePack> packs = new ArrayList<>();
+	public static LoaderComplexFabric INSTANCE;
+
+	public LoaderComplexFabric() {
+		super();
+		loader = new FabricLoader();
+		resourceHelper = mod -> packs.add( new FabricResourcePack(mod) );
+		INSTANCE = this;
+	}
+
+	@Override
+	public void onInitialize() {
+		initAddons();
+	}
+
+	public Loader getLoader() {
+		return loader;
+	}
+
+}
