+++ loadercomplex-fabric-12/src/main/java/com/enderzombi102/loadercomplex/fabric12/impl/entity/FabricItemEntity.java
@@ -1,0 +1,24 @@
+package com.enderzombi102.loadercomplex.fabric12.impl.entity;
+
+import com.enderzombi102.loadercomplex.fabric12.impl.item.FabricItemStack;
+import com.enderzombi102.loadercomplex.api.entity.ItemEntity;
+import com.enderzombi102.loadercomplex.api.item.ItemStack;
+
+public class FabricItemEntity extends FabricEntity implements ItemEntity {
+	private final net.minecraft.entity.ItemEntity backingEntity;
+
+	public FabricItemEntity(net.minecraft.entity.ItemEntity backingEntity) {
+		super( backingEntity );
+		this.backingEntity = backingEntity;
+	}
+
+	@Override
+	public ItemStack getStack() {
+		return new FabricItemStack( this.backingEntity.getItemStack() );
+	}
+
+	@Override
+	public void setStack(ItemStack stack) {
+		this.backingEntity.setItemStack( (net.minecraft.item.ItemStack) stack.getStack() );
+	}
+}
