+++ loadercomplex-fabric-12/src/main/resources/fabric.mod.json
@@ -23,7 +23,8 @@
 	},
 
 	"mixins": [
-		"loadercomplex-fabric.mixins.json"
+		"loadercomplex-fabric12.mixins.json",
+		"loadercomplex-fabric17.mixins.json"
 	],
 
 	"depends": {
