--- loadercomplex-fabric-12/src/main/resources/loadercomplex-fabric.mixins.json
@@ -1,18 +1,0 @@
-{
-	"required": true,
-	"minVersion": "0.8",
-	"package": "com.enderzombi102.loaderComplex.fabric12.mixin",
-	"compatibilityLevel": "JAVA_8",
-	"mixins": [
-		"ItemMixin",
-		"ReloadableResourceManagerImplMixin",
-		"SimpleRegistryMixin"
-	],
-	"client": [
-		"MinecraftClientMixin",
-		"ModelLoaderMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
